tests: restructure tests dir to match filesystem structure
